desafio completo

+++ src/services/product.service.js
@@ -1,6 +1,7 @@
 import { productDto } from "../dto/product-response.dto.js";
 import productRepository from "../persistences/mongo/repositories/product.repository.js";
 import error from "../errors/customsErrors.js"
+import { logger } from "../utils/logger.js";
 
 const getAllProducts = async (query, options) => {
     const products = await productRepository.getAllProducts(query, options);
@@ -15,9 +16,19 @@ const getProductById = async (id) => {
     return product;
 }
 
-const createProduct = async (data) => {
-    const product = await productRepository.createProduct(data);
-    if(!product) throw error.missingInfo(msg);
+// const createProduct = async (data) => {
+//     const product = await productRepository.createProduct(data);
+//     if(!product) throw error.missingInfo(msg);
+//     return product;
+// }
+const createProduct = async (data, user) => {
+    let productData = data;
+    //logger.info(user.role)
+    if(user.role === "premium"){
+        productData = {...data, owner:user._id}
+        //logger.info(productData);
+    }
+    const product = await productRepository.createProduct(productData);
     return product;
 }
 
@@ -26,7 +37,11 @@ const updateProduct =  async (id, data) => {
     return product;
 }
 
-const deleteProduct = async (id) => {
+const deleteProduct = async (id, user) => {
+    const productData = await productRepository.getProductById (id);
+    if(user.role === "premium" && productData.owner !== user._id ){
+        throw error.unauthorizedError("Usuario no autorizado");
+    }
     const product = await productRepository.deleteProduct(id);
     return product;
 }
